+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add a `versions` field to `cmd` to specify on which versions of recipe to run a command
 - Add a way to specify a base recipe with `from: base-recipe`. More information in the docs
 - Add `--no-color` option that disables colored output
+- Add `--output-dir` flag to `build` subcommand to override output directory for this build
 
 # 0.8.0
 - Dependencies defined for default images like `pkger-deb` or `pkger-rpm` will be used for custom images using the same output target. This means that all images used for building RPM will use dependencies defined for `pkger-rpm` - cleaner recipes ;)